+++ src/carlson/elliprc.rs
@@ -86,6 +86,10 @@ mod tests {
 
     #[test]
     fn test_elliprc() {
-        compare_test_data!("./tests/data/boost/elliprc_data.txt", _elliprc, 2.2e-16);
+        compare_test_data!(
+            "./tests/data/boost/elliprc_data.txt",
+            _elliprc,
+            f64::EPSILON
+        );
     }
 }